refactoring code javascript and I add loader of sweetalert and I show messages of errors in login
