+++ resources/views/auth/login.blade.php
@@ -12,42 +12,42 @@
                 <p class="text-muted">Ingrese a su cuenta</p>
                 <form method="POST" action="{{ route('login') }}">
                     @csrf
-                    <div class="input-group mb-3">
-                    <div class="input-group-prepend">
-                        <span class="input-group-text">
-                        <i class="fa fa-user"></i>
-                        </span>
-                    </div>
-                    <input
-                        class="form-control"
-                        type="text"
-                        placeholder="Correo electrónico"
-                        name="email"
-                        value="{{ old('email') }}"
-                    >
                     @error('email')
-                    <span class="invalid-feedback" role="alert">
+                    <span class="error text-danger">
                         <strong>{{ $message }}</strong>
                     </span>
                     @enderror
+                    <div class="input-group mb-3">
+                        <div class="input-group-prepend">
+                            <span class="input-group-text">
+                            <i class="fa fa-user"></i>
+                            </span>
+                        </div>
+                        <input
+                            class="form-control"
+                            type="text"
+                            placeholder="Correo electrónico"
+                            name="email"
+                            value="{{ old('email') }}"
+                            >
                     </div>
-                    <div class="input-group mb-4">
-                    <div class="input-group-prepend">
-                        <span class="input-group-text">
-                        <i class="fa fa-key"></i>
-                        </span>
-                    </div>
-                    <input
-                        class="form-control"
-                        type="password"
-                        placeholder="Contraseña"
-                        name="password"
-                    >
                     @error('password')
-                        <span class="invalid-feedback" role="alert">
+                        <span class="error text-danger" role="alert">
                             <strong>{{ $message }}</strong>
                         </span>
                     @enderror
+                    <div class="input-group mb-4">
+                        <div class="input-group-prepend">
+                            <span class="input-group-text">
+                            <i class="fa fa-key"></i>
+                            </span>
+                        </div>
+                        <input
+                            class="form-control"
+                            type="password"
+                            placeholder="Contraseña"
+                            name="password"
+                        >
                     </div>
                     <div class="row">
                         <div class="col-6">
