+++ resources/views/categories/index.blade.php
@@ -22,183 +22,14 @@
     </div>
 
     <!-- Create Modal -->
-    <div class="modal fade" id="modalCategory" data-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
-        <div class="modal-dialog">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="staticBackdropLabel">Nueva Categoria</h5>
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-                </div>
-                <div class="modal-body">
-                    <form id="form-category" enctype="multipart/form-data">
-                        <input type="hidden" name="_token" id="token" value="{{ csrf_token() }}">
-                        <div class="row">
-                            <div class="col-md-12">
-                                <div class="form-group">
-                                    <label for="name">Nombre de categoria</label>
-                                    <input
-                                        type="text"
-                                        id="name"
-                                        name="name"
-                                        placeholder="Ingrese nombre de categoria"
-                                        class="form-control">
-                                </div>
-                            </div>
-                        </div>
-                        <div class="row">
-                            <div class="col-md-12">
-                                <div class="form-group">
-                                    <label for="status">Estado</label>
-                                    <select name="status" id="status" class="form-control">
-                                        <option value="0" selected disabled>-- Seleccione --</option>
-                                        <option value="A">Activo</option>
-                                        <option value="I">Inactivo</option>
-                                    </select>
-                                </div>
-                            </div>
-                        </div>
-                        <div class="row">
-                            <div class="col-md-12">
-                                <div class="form-group">
-                                    <p><b>Imagen</b></p>
-                                    <div class="input-group mb-3">
-                                        <div class="input-group-prepend">
-                                          <span class="input-group-text" id="inputGroupFileAddon01">Subir</span>
-                                        </div>
-                                        <div class="custom-file">
-                                            <input type="file" class="custom-file-input" id="image_category" aria-describedby="inputGroupFileAddon01">
-                                            <label id="img_category" class="custom-file-label" for="image_category">Escoger archivo</label>
-                                        </div>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </form>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" data-dismiss="modal">Cancelar</button>
-                    <button id="save-button" type="button" class="btn btn-primary">Guardar</button>
-                </div>
-            </div>
-        </div>
-    </div>
-    <!-- Edit Modal -->
-    <div class="modal fade" id="modalCategoryEdit" data-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
-        <div class="modal-dialog">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="staticBackdropLabel">Editar Categoria</h5>
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-                </div>
-                <div class="modal-body">
-                    <form id="form-category-edit">
-                        <input type="hidden" name="_token" id="token" value="{{ csrf_token() }}">
-                        <div class="row">
-                            <div class="col-md-12">
-                                <div class="form-group">
-                                    <label for="name-edit">Nombre de categoria</label>
-                                    <input
-                                        type="text"
-                                        id="name"
-                                        name="name"
-                                        placeholder="Ingrese nombre de categoria"
-                                        value=""
-                                        class="form-control">
-                                </div>
-                            </div>
-                        </div>
-                        <div class="row">
-                            <div class="col-md-12">
-                                <div class="form-group">
-                                    <label for="status-edit">Estado</label>
-                                    <select name="status" id="status-edit" class="form-control">
-                                        <option value="0" selected disabled>-- Seleccione --</option>
-                                        <option value="A">Activo</option>
-                                        <option value="I">Inactivo</option>
-                                    </select>
-                                </div>
-                            </div>
-                        </div>
-                    </form>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" data-dismiss="modal">Cancelar</button>
-                    <button id="edit-button" type="button" class="btn btn-primary">Editar</button>
-                </div>
-            </div>
-        </div>
-    </div>
+
+    @include('categories.modalCreate')
+
+    <!-- Edit Modal Text -->
+    @include('categories.modalEditText')
 
     <!-- Edit Modal Image-->
-    <div class="modal fade" id="modalImageEdit" data-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
-        <div class="modal-dialog">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="staticBackdropLabel">Cambiar Imagen de Categoria</h5>
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-                </div>
-                <div class="modal-body">
-                    <form id="form-category-edit-image" enctype="multipart/form-data">
-                        <input type="hidden" name="_token" id="token" value="{{ csrf_token() }}">
-                        <div class="row">
-                            <div class="col-md-12">
-                                <p><b>Imagen</b></p>
-                                <img src="" id="img-category" class="img-fluid" alt="">
-                            </div>
-                        </div>
-                        <div class="row mt-5">
-                            <div class="col-md-12">
-                                <div class="form-group">
-                                    <div class="input-group mb-3">
-                                        <div class="input-group-prepend">
-                                          <span class="input-group-text" id="inputGroupFileAddon01">Subir</span>
-                                        </div>
-                                        <div class="custom-file">
-                                            <input type="file" class="custom-file-input" id="image_category_edit" aria-describedby="inputGroupFileAddon01">
-                                            <label id="img_category_edit" class="custom-file-label" for="image_category_edit">Escoger archivo</label>
-                                        </div>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </form>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" data-dismiss="modal">Cancelar</button>
-                    <button id="change-image-button" type="button" class="btn btn-primary">Editar</button>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <div class="modal fade" id="modalDelete" data-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
-        <div class="modal-dialog">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="staticBackdropLabel">¿Seguro que desea eliminar?</h5>
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-                </div>
-                <div class="modal-body">
-                    <form id="delete-category">
-                        <input type="hidden" name="_token" id="token" value="{{ csrf_token() }}">
-                    </form>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" data-dismiss="modal">Cancelar</button>
-                    <button id="delete-button" type="button" class="btn btn-primary">Eliminar</button>
-                </div>
-            </div>
-        </div>
-    </div>
-
+    @include('categories.modalEditImage')
 
     <table id="table-category" class="display nowrap table table-bordered table-hover" style="width: 100%;">
         <thead>
@@ -258,7 +89,5 @@
         ]
     });
 </script>
-<script src="{{ asset('assets/js/categories/create-category.js') }}"></script>
-<script src="{{ asset('assets/js/categories/edit-category.js') }}"></script>
-<script src="{{ asset('assets/js/categories/delete-category.js') }}"></script>
+<script src="{{ asset('assets/js/categories/categories.js') }}" type="module"></script>
 @endpush
